TIP-458: Move MediaPathTransformer
