+++ src/Pim/Component/Connector/Reader/File/MediaPathTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\Connector\Reader\File\Product;
+namespace Pim\Component\Connector\Reader\File;
 
 use Pim\Component\Catalog\Repository\AttributeRepositoryInterface;
