+++ src/Pim/Component/Connector/Reader/File/Product/CsvProductReader.php
@@ -5,6 +5,7 @@ namespace Pim\Component\Connector\Reader\File\Product;
 use Pim\Component\Connector\ArrayConverter\ArrayConverterInterface;
 use Pim\Component\Connector\Reader\File\CsvReader;
 use Pim\Component\Connector\Reader\File\FileIteratorFactory;
+use Pim\Component\Connector\Reader\File\MediaPathTransformer;
 
 /**
  * Product csv reader
