+++ src/Pim/Component/Connector/spec/Reader/File/Product/XlsxProductReaderSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Component\Connector\Reader\File\Product;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Connector\Reader\File\FileIterator;
 use Pim\Component\Connector\Reader\File\FileIteratorInterface;
-use Pim\Component\Connector\Reader\File\Product\MediaPathTransformer;
+use Pim\Component\Connector\Reader\File\MediaPathTransformer;
 use Prophecy\Argument;
 
 class CsvProductReaderSpec extends ObjectBehavior
